requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cycler==0.12.1
debugpy==1.8.16
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.116.2
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.30.1
joblib==1.5.2
jsonschema==4.25.1
jupyter_client==8.6.3
jupyter_core==5.8.1
librosa==0.11.0
llvmlite==0.44.0
loralib==0.1.2
matplotlib==3.10.6
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
openai-whisper==20250625
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
psutil==7.0.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
seaborn==0.13.2
sentence-transformers==5.1.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==2.3.7
